app/requirements.txt
SQLAlchemy==2.0.23
pymysql==1.1.0
Faker==19.6.2
click==8.1.7